--- utag/__init__.py
@@ -1,21 +1,0 @@
-import warnings
-
-from scipy.sparse import SparseEfficiencyWarning
-
-from utag.segmentation import utag
-
-
-try:
-    # Even though there is no "imc/_version" file,
-    # it should be generated by
-    # setuptools_scm when building the package
-    from utag._version import version
-
-    __version__ = version
-except ImportError:
-    from setuptools_scm import get_version as _get_version
-
-    version = __version__ = _get_version(root="..", relative_to=__file__)
-
-warnings.simplefilter("ignore", FutureWarning)
-warnings.simplefilter("ignore", SparseEfficiencyWarning)
